fibonacci to bigint

+++ src/main/scala-2.12/algorithms/fibonacci.scala
@@ -1,9 +1,9 @@
 package algorithms
 
 object fibonacci {
-  def fib(n:Int): Int = {
+  def fib(n: Int): BigInt = {
     @annotation.tailrec
-    def loop(togo: Int, current:Int=1, acc:Int=0):Int = {
+    def loop(togo: Int, current: BigInt = 1, acc: BigInt = 0): BigInt = {
       if(togo == 1) acc
       else loop(togo-1, acc+current, current)
     }
